refactor: Make studentCardUpload template
