+++ src/components/shared/Templates/StudentCardUpload/style.module.scss
