+++ src/components/shared/Templates/index.ts
@@ -1,2 +1,3 @@
 export { default as FriendsList } from './FriendsList';
 export { default as MeetingApplyTemplate } from './MeetingApplyTemplate';
+export { default as StudentCardUpload } from './StudentCardUpload';
